// src/adapter/coins/btc/dto/btc-address-generate-response.dto.ts
/**
 * Adapter response payload for BTC address generation.
 */
export interface BtcAddressGenerateAdapterResponse {
  address: string;
  keys: {
    public: string;
    private: string;
  };
  derivation: {
    path: string;
    purpose: number;
    coin: number;
    account: number;
    change: number;
    index: number;
  };
}